+++ CHANGELOG.md
@@ -2,12 +2,19 @@
 
 ## Current (in progress)
 
+- Return HTTP 410 Gone response for deleted resources [#56](https://github.com/datagouv/api-tabular/pull/56)
+
+## 0.2.5 (2025-07-21)
+
 - Remove PostgREST version in health endpoint [#43](https://github.com/datagouv/api-tabular/pull/43)
 - Better health endpoint [#45](https://github.com/datagouv/api-tabular/pull/45) [#46](https://github.com/datagouv/api-tabular/pull/46)
 - Use indexes from resources_exceptions table to allow operations or not [#44](https://github.com/datagouv/api-tabular/pull/44)
 - Refactor tests to use the test containers [#48](https://github.com/datagouv/api-tabular/pull/48)
-- Use a standard `pyproject.toml` file, use Poetry 2, use a lightweight image with integrated Poetry 2 for linting an build CI jobs, and don't wait for install step before running the tests jobs in CI [#38](https://github.com/datagouv/api-tabular/pull/38)
+- Use a standard `pyproject.toml` file, use Poetry 2, use a lightweight image with integrated Poetry 2 for linting an build CI jobs, and don't wait for install step before running the tests jobs in CI [#38](https://github.com/datagouv/api-tabular/pull/38) [#54](https://github.com/datagouv/api-tabular/pull/54)
 - Add endpoint to see aggregation exceptions [#47](https://github.com/datagouv/api-tabular/pull/47)
+- Add endpoint to get data as JSON [#49](https://github.com/datagouv/api-tabular/pull/49)
+- Use PostgreSQL 15 for containerized test DB to be iso with prod [#51](https://github.com/datagouv/api-tabular/pull/51)
+- Improve the documentation [#52](https://github.com/datagouv/api-tabular/pull/52) and [#53](https://github.com/datagouv/api-tabular/pull/53)
 
 ## 0.2.4 (2025-03-20)